+++ src/pages/review/[reviewId]/index.tsx
@@ -17,6 +17,7 @@ import { useViewport } from '@/hooks/useViewport';
 import _ from 'lodash';
 import { RealtimeChannel } from '@supabase/supabase-js';
 import { supabase } from '@/libs/supabase';
+import { useAlarmSubscribeComment } from '@/hooks/useAlarmSubscribeComment';
 
 const ReviewPage = () => {
   const router = useRouter();
@@ -48,35 +49,8 @@ const ReviewPage = () => {
 
   const placeId = review?.place_id;
 
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    if (!currentUserId) return;
-    const subscription: RealtimeChannel = supabase
-      .channel('custom-insert-channel')
-      .on(
-        'postgres_changes',
-        { event: 'INSERT', schema: 'public', table: 'comments' },
-        (payload) => {
-          queryClient.invalidateQueries({
-            queryKey: ['comments', reviewId],
-          });
-        },
-      )
-      .on(
-        'postgres_changes',
-        { event: 'DELETE', schema: 'public', table: 'comments' },
-        (payload) => {
-          queryClient.invalidateQueries({
-            queryKey: ['comments', reviewId],
-          });
-        },
-      )
-      .subscribe();
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, []);
+  // 댓글 실시간
+  useAlarmSubscribeComment(currentUserId, reviewId);
 
   if (isLoading) {
     return (
